refactor modeling code
